# Increment/Decrement Valid Number Check