--- a/libs/server/Resp/Objects/ListCommands.cs
+++ b/libs/server/Resp/Objects/ListCommands.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Linq;
-using System.Text;
 using Garnet.common;
 using Tsavorite.core;