--- a/libs/server/TLS/IGarnetTlsOptions.cs
+++ b/libs/server/TLS/IGarnetTlsOptions.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-using System;
 using System.Net.Security;
 
 namespace Garnet.server.TLS