Added validity check to catch increment/decrement of alphanumeric strings.
This is a proposed solution that minimizes the perf overhead of passing the error from RMWMethods.
However, the overhead of TryBytesToLong will not be avoided.
Maybe it is possible to include a type at upsert and avoid TryBytesToLong validation every time INCR/DECR is executed.

Second iteration updated the PR to catch overflow exceptions and to validate operands on copy update and at InitialUpdate.

Third iteration combined bdn.benchmark with embedded bench to evaluate performance of new validity checks.

**Main without validity checks**

| Method      | Job    | EnvironmentVariables | Runtime  | input    | Mean     | Error   | StdDev  |
|------------ |------- |--------------------- |--------- |--------- |---------:|--------:|--------:|
| Increment   | .NET 6 | Empty                | .NET 6.0 | ?        | 297.4 ns | 0.51 ns | 0.47 ns |
| Decrement   | .NET 6 | Empty                | .NET 6.0 | ?        | 303.4 ns | 0.49 ns | 0.38 ns |
| Increment   | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | ?        | 206.5 ns | 0.28 ns | 0.25 ns |
| Decrement   | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | ?        | 214.2 ns | 0.55 ns | 0.49 ns |
| IncrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[30] | 210.0 ns | 0.34 ns | 0.32 ns |
| DecrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[30] | 214.8 ns | 0.30 ns | 0.28 ns |
| IncrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[30] | 132.7 ns | 0.30 ns | 0.29 ns |
| DecrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[30] | 131.2 ns | 0.36 ns | 0.34 ns |
| IncrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[31] | 227.9 ns | 1.09 ns | 1.02 ns |
| DecrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[31] | 236.1 ns | 0.86 ns | 0.81 ns |
| IncrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[31] | 146.0 ns | 0.27 ns | 0.25 ns |
| DecrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[31] | 145.0 ns | 0.50 ns | 0.44 ns |
| IncrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[40] | 240.3 ns | 0.80 ns | 0.75 ns |
| DecrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[40] | 242.9 ns | 0.62 ns | 0.58 ns |
| IncrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[40] | 157.1 ns | 0.17 ns | 0.16 ns |
| DecrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[40] | 156.7 ns | 0.25 ns | 0.23 ns |
| IncrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[41] | 234.3 ns | 1.43 ns | 1.34 ns |
| DecrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[41] | 244.8 ns | 0.78 ns | 0.73 ns |
| IncrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[41] | 158.0 ns | 0.83 ns | 0.77 ns |
| DecrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[41] | 161.9 ns | 0.48 ns | 0.45 ns |


**PR with validity checks**

| Method      | Job    | EnvironmentVariables | Runtime  | input    | Mean     | Error   | StdDev  |
|------------ |------- |--------------------- |--------- |--------- |---------:|--------:|--------:|
| Increment   | .NET 6 | Empty                | .NET 6.0 | ?        | 327.6 ns | 0.90 ns | 0.84 ns |
| Decrement   | .NET 6 | Empty                | .NET 6.0 | ?        | 259.3 ns | 1.09 ns | 1.02 ns |
| Increment   | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | ?        | 232.5 ns | 1.10 ns | 1.03 ns |
| Decrement   | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | ?        | 175.2 ns | 0.28 ns | 0.26 ns |
| IncrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[30] | 201.1 ns | 1.03 ns | 0.97 ns |
| DecrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[30] | 200.0 ns | 0.29 ns | 0.26 ns |
| IncrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[30] | 135.1 ns | 0.27 ns | 0.25 ns |
| DecrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[30] | 134.6 ns | 0.13 ns | 0.12 ns |
| IncrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[31] | 226.6 ns | 0.16 ns | 0.13 ns |
| DecrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[31] | 234.0 ns | 1.32 ns | 1.24 ns |
| IncrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[31] | 150.1 ns | 0.22 ns | 0.19 ns |
| DecrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[31] | 153.0 ns | 0.38 ns | 0.34 ns |
| IncrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[40] | 244.4 ns | 1.61 ns | 1.50 ns |
| DecrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[40] | 248.5 ns | 2.54 ns | 2.38 ns |
| IncrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[40] | 163.0 ns | 0.30 ns | 0.26 ns |
| DecrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[40] | 162.9 ns | 0.15 ns | 0.14 ns |
| IncrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[41] | 252.9 ns | 1.95 ns | 1.83 ns |
| DecrementBy | .NET 6 | Empty                | .NET 6.0 | Byte[41] | 242.2 ns | 2.31 ns | 2.16 ns |
| IncrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[41] | 164.5 ns | 0.35 ns | 0.32 ns |
| DecrementBy | .NET 8 | DOTNET_TieredPGO=0   | .NET 8.0 | Byte[41] | 163.2 ns | 0.26 ns | 0.22 ns |